needforspeed: use Py_ssize_t for the fastsearch counter and skip length (thanks, neal!).  and yes, I've verified that this doesn't slow things down ;-)

+++ Objects/stringobject.c
@@ -802,7 +802,7 @@ LOCAL(Py_ssize_t)
 fastsearch(const char* s, Py_ssize_t n, const char* p, Py_ssize_t m, int mode)
 {
 	long mask;
-	int skip, count = 0;
+	Py_ssize_t skip, count = 0;
 	Py_ssize_t i, j, mlast, w;
 
 	w = n - m;
